01_demo_s3/requirements.txt
flask==2.3.3
boto3==1.28.85
python-dotenv==1.0.0